Snake game formatting and added input + rewards
Updated the collision input values that check if there's a wall/body in each direction of the snake (besides directly behind it since that's the body) to a distance_to_collision function which checks how far the snake has in each direction before hitting something that'll get a game over. Also put the distance to fruit number in the state array. All values are normalized between 0 and 1.

Reorganized the code by making a get_reward function that calls all the things that give rewards just so it's a little easier to follow.