Removed risk bonus and distance reward
Removed the reward that gave a bonus for getting fruits that are in risk places, and also removed the reward for getting closer or further away to the fruit. This is because I noticed the snake did better organizing itself in tight spaces (when the distance reward was automatically turned off). Also updated gradient clipping to 10. This marked a new high for the weighted score of 816 (previous high was 795)